relayered the code to work with github vulndb worklflow, cleaned up export code and fixed somem minor bugs
