+++ cmd/devguard-cli/commands/vulndb.go
@@ -1,21 +1,14 @@
 package commands
 
 import (
-	"context"
-	"encoding/csv"
 	"errors"
-	"fmt"
-	"log"
 	"log/slog"
 	"os"
-	"path/filepath"
 	"regexp"
 	"slices"
-	"strconv"
 	"strings"
 	"time"
 
-	"github.com/jackc/pgx/v5/pgxpool"
 	"github.com/l3montree-dev/devguard/internal/core"
 	"github.com/l3montree-dev/devguard/internal/core/vulndb"
 	"github.com/l3montree-dev/devguard/internal/database"
@@ -146,83 +139,24 @@ func newImportCommand() *cobra.Command {
 		Short: "Will import the vulnerability database",
 		Args:  cobra.MaximumNArgs(2),
 		Run: func(cmd *cobra.Command, args []string) {
-			if len(args) == 0 {
-				slog.Error("missing mode choose <inc> for incremental updates (recommended) or <full> to copy the full vuln database")
-				return
-			}
-			mode := args[0]
 			core.LoadConfig() // nolint
 
+			var mode string
+			if len(args) > 0 {
+				mode = args[0]
+			}
+			// import incremental updates using the diff files
 			if mode == "inc" {
-				username := os.Getenv("POSTGRES_USER")
-				password := os.Getenv("POSTGRES_PASSWORD")
-				host := os.Getenv("POSTGRES_HOST")
-				port := os.Getenv("POSTGRES_PORT")
-				dbname := os.Getenv("POSTGRES_DB")
-
-				// replace with your PostgreSQL connection string
-				connStr := fmt.Sprintf("postgres://%s:%s@%s:%s/%s", username, password, host, port, dbname)
-				// create a connection pool with increased connections for parallel processing
-				ctx := context.Background()
-				config, err := pgxpool.ParseConfig(connStr)
+				err := vulndb.ImportFromDiff()
 				if err != nil {
-					log.Fatalf("Unable to parse config: %v", err)
-				}
-				// increase pool size for parallel operations
-				config.MaxConns = 10
-				config.MinConns = 2
-
-				pool, err := pgxpool.NewWithConfig(ctx, config)
-
-				if err != nil {
-					log.Fatalf("Unable to create connection pool: %v", err)
-				}
-				defer pool.Close()
-				// only import the incremental updates
-				// pull incremental files from github database
-				files, err := os.ReadDir("vulndb-tmp")
-				if err != nil {
-					slog.Error("error when reading dir", "error", err)
+					slog.Error("could not import from diff files")
 					return
 				}
-
-				for _, file := range files {
-					name := file.Name()
-					if filepath.Ext(name) != ".csv" {
-						continue
-					}
-					name = strings.TrimRight(name, ".csv")
-					fields := strings.Split(name, "_")
-					if len(fields) != 3 || fields[1] != "diff" {
-						continue
-					}
-					mode := fields[2]
-					table := fields[0]
-					switch mode {
-					case "insert":
-						slog.Info("start inserting", "file", name)
-						err = processInsertDiff(ctx, pool, "vulndb-tmp/"+name+".csv", table)
-						if err != nil {
-							continue
-						}
-					case "delete":
-						slog.Info("start deleting", "file", name)
-						err = processDeleteDiff(ctx, pool, "vulndb-tmp/"+name+".csv", table)
-						if err != nil {
-							continue
-						}
-					case "update":
-						slog.Info("start updating", "file", name)
-						err = processUpdateDiff(ctx, pool, "vulndb-tmp/"+name+".csv", table)
-						if err != nil {
-							continue
-						}
-					default:
-						slog.Warn("invalid mode for diff file")
-					}
+			} else { // import the data base from the package master
+				if mode == "diff" { // additionally create a diff table to be used by the export command
+					os.Setenv("MAKE_DIFF_TABLES", "true")
 				}
-				return
-			} else if mode == "full" { // import the full table
+				// import the full table
 				database, err := core.DatabaseFactory()
 				if err != nil {
 					slog.Error("could not connect to database", "error", err)
@@ -245,136 +179,11 @@ func newImportCommand() *cobra.Command {
 					slog.Error("could not import vulndb", "err", err)
 					return
 				}
-			} else { // invalid argument
-				slog.Error("invalid first argument, most be 'inc' to only import the difference or 'full' to import the whole table")
-				return
 			}
-			return
+
 		},
 	}
 	return importCmd
-}
-
-func processInsertDiff(ctx context.Context, pool *pgxpool.Pool, filePath string, tableName string) error {
-	if tableName != "cves" {
-		return nil
-	}
-	fd, err := os.Open(filePath)
-	if err != nil {
-		return err
-	}
-	defer fd.Close()
-	conn, err := pool.Acquire(ctx)
-	if err != nil {
-		return err
-	}
-	defer conn.Conn().Close(ctx)
-	slog.Info("start copying")
-	result, err := conn.Conn().PgConn().CopyFrom(ctx, fd, fmt.Sprintf("COPY %s FROM STDIN WITH (FORMAT csv, HEADER true, NULL 'NULL')", tableName))
-	if err != nil {
-		slog.Error("TOT", "err", err, "result", result)
-		return err
-	}
-	slog.Info("finished copying")
-
-	return nil
-}
-
-func processDeleteDiff(ctx context.Context, pool *pgxpool.Pool, filePath string, tableName string) error {
-	if tableName != "cves" {
-		return nil
-	}
-	fd, err := os.Open(filePath)
-	if err != nil {
-		return err
-	}
-	defer fd.Close()
-	conn, err := pool.Acquire(ctx)
-	if err != nil {
-		return err
-	}
-	defer conn.Conn().Close(ctx)
-
-	csvReader := csv.NewReader(fd)
-	allRecords, err := csvReader.ReadAll()
-	if err != nil {
-		return err
-	}
-
-	primaryKey := allRecords[0][0]
-	allRecords = allRecords[1:]
-	slog.Info("start deleting")
-	for i := range allRecords {
-		key := allRecords[i][0]
-		sql := fmt.Sprintf("DELETE FROM %s WHERE %s.%s = %s", tableName, tableName, primaryKey, "'"+key+"'")
-		result, err := conn.Exec(ctx, sql)
-		if err != nil {
-			slog.Error("TOT", "err", err, "result", result)
-			continue
-		}
-	}
-	slog.Info("finished deleting")
-
-	return nil
-}
-
-func processUpdateDiff(ctx context.Context, pool *pgxpool.Pool, filePath string, tableName string) error {
-	if tableName != "cves" {
-		return nil
-	}
-	fd, err := os.Open(filePath)
-	if err != nil {
-		return err
-	}
-	defer fd.Close()
-	conn, err := pool.Acquire(ctx)
-	if err != nil {
-		return err
-	}
-	defer conn.Conn().Close(ctx)
-
-	csvReader := csv.NewReader(fd)
-	record, err := csvReader.Read()
-	if err != nil {
-		return err
-	}
-	columnsToUpdate := record[1:] // exclude primary key(s)
-	for i, column := range columnsToUpdate {
-		if column == "references" {
-			columnsToUpdate[i] = fmt.Sprintf("\"%s\" = EXCLUDED.%s", column, column)
-		} else {
-			columnsToUpdate[i] = fmt.Sprintf("%s = EXCLUDED.%s", column, column)
-		}
-	}
-	assignSQL := strings.Join(columnsToUpdate, ", ")
-
-	tmpTable := tableName + "_tmp_" + strconv.Itoa(time.Now().Second())
-
-	_, err = conn.Conn().Exec(ctx, fmt.Sprintf("CREATE TABLE %s (LIKE %s INCLUDING ALL);", tmpTable, tableName))
-	if err != nil {
-		return fmt.Errorf("failed to create tmp table: %w", err)
-	}
-	defer conn.Conn().Exec(ctx, fmt.Sprintf("DROP TABLE %s;", tmpTable))
-	fd, err = os.Open(filePath)
-	if err != nil {
-		return err
-	}
-
-	_, err = conn.Conn().PgConn().CopyFrom(ctx, fd, fmt.Sprintf("COPY %s FROM STDIN WITH (FORMAT csv, HEADER true, NULL 'NULL')", tmpTable))
-	if err != nil {
-		return fmt.Errorf("failed to copy to tmp table: %w", err)
-	}
-
-	pkeys := primaryKeysFromTables[tableName]
-
-	upsertSQL := fmt.Sprintf("INSERT INTO %s SELECT * FROM %s ON CONFLICT (%s) DO UPDATE SET %s", tableName, tmpTable, pkeys[0], assignSQL)
-
-	if _, err := conn.Exec(ctx, upsertSQL); err != nil {
-		return err
-	}
-	slog.Info("update completed")
-
-	return nil
 }
 
 func newSyncCommand() *cobra.Command {
@@ -524,28 +333,9 @@ func newExportIncrementalCommand() *cobra.Command {
 		Run: func(cmd *cobra.Command, args []string) {
 			// first import the new state
 			core.LoadConfig() // nolint
-
-			database, err := core.DatabaseFactory()
+			os.RemoveAll("diffs-tmp/")
+			err := vulndb.Export()
 			if err != nil {
-				slog.Error("could not connect to database", "err", err)
-				return
-			}
-			migrateDB(database)
-
-			cveRepository := repositories.NewCVERepository(database)
-			cweRepository := repositories.NewCWERepository(database)
-			exploitsRepository := repositories.NewExploitRepository(database)
-			affectedComponentsRepository := repositories.NewAffectedComponentRepository(database)
-
-			tag := "latest"
-			if len(args) > 0 {
-				tag = args[0]
-			}
-			os.Setenv("MAKE_TABLE_DIFF", "true")
-			v := vulndb.NewImportService(cveRepository, cweRepository, exploitsRepository, affectedComponentsRepository)
-			err = v.Import(database, tag)
-			if err != nil {
-				slog.Error("could not import vulndb", "err", err)
 				return
 			}
 		},
